Remove tool's orgToolObject from state
