+++ src/app/shared/state/my-entries.query.ts
@@ -1,6 +1,5 @@
 import { Injectable } from '@angular/core';
 import { Query } from '@datorama/akita';
-import { OrgToolObject } from 'app/mytools/my-tool/my-tool.component';
 import { OrgWorkflowObject } from 'app/myworkflows/my-workflow/my-workflow.component';
 import { Observable } from 'rxjs';
 import { map } from 'rxjs/operators';
@@ -8,7 +7,7 @@ import { MyEntriesState, MyEntriesStore } from './my-entries.store';
 
 @Injectable()
 export class MyEntriesQuery extends Query<MyEntriesState> {
-  groupEntriesObject$: Observable<Array<OrgToolObject | OrgWorkflowObject>> = this.select(state => state.groupEntriesObject);
+  groupEntriesObject$: Observable<Array<OrgWorkflowObject>> = this.select(state => state.groupEntriesObject);
   hasGroupEntriesObject$: Observable<boolean> = this.groupEntriesObject$.pipe(
     map(groupEntriesObject => {
       return groupEntriesObject && groupEntriesObject.length !== 0;
