+++ src/app/shared/state/my-entries.store.ts
@@ -1,11 +1,10 @@
 import { Injectable } from '@angular/core';
 import { Store, StoreConfig } from '@datorama/akita';
-import { OrgToolObject } from 'app/mytools/my-tool/my-tool.component';
 import { OrgWorkflowObject } from 'app/myworkflows/my-workflow/my-workflow.component';
 
 export interface MyEntriesState {
   refreshingMyEntries: boolean;
-  groupEntriesObject: Array<OrgToolObject | OrgWorkflowObject>;
+  groupEntriesObject: Array<OrgWorkflowObject>;
 }
 
 export function createInitialState(): MyEntriesState {
